+++ src/app/survey/types/survey.ts
