--- src/app/surveys/page.tsx
@@ -1,9 +1,0 @@
-import SurveyForm from "@/components/templates/SurveyForm";
-
-export default function SurveysPage() {
-    return (
-        <div className={"h-screen flex"}>
-            <SurveyForm/>
-        </div>
-    );
-}
